+++ packages/react/src/core/components/create-component.tsx
@@ -138,7 +138,7 @@ function getSlotName<Y extends string>(slot?: ComponentSlot<Y>) {
   if (isArray(slot))
     return slot.map((value) => toPascalCase(value as string)).join("")
   else if (isObject(slot)) return toPascalCase(slot.name)
-  else return toPascalCase(slot as string)
+  else return toPascalCase(slot)
 }
 
 export function createComponent<
@@ -413,9 +413,7 @@ export function createSlotComponent<
     T extends keyof R = keyof R,
   >(
     el: FC<WithoutThemeProps<R, M, T>> | H,
-    slot: ComponentSlot<ComponentSlotName<M>> = "root" as ComponentSlot<
-      ComponentSlotName<M>
-    >,
+    slot: ComponentSlot<ComponentSlotName<M>> = "root",
     {
       name,
       className,